+++ test/integration/test_hook_writer.rb
@@ -22,7 +22,7 @@ class TestHookWriter < Minitest::Test
   def test_should_write_the_hook_s_script
     in_git_dir { install }
 
-    assert_equal FunCi::Setup::HookScript.for("pre-commit"), File.read(hook_path)
+    assert_equal FunCi::Setup::HookScript.for("post-commit"), File.read(hook_path)
   end
 
   def test_should_make_the_hook_executable
@@ -40,7 +40,7 @@ class TestHookWriter < Minitest::Test
   def test_should_say_which_hook_it_installed
     in_git_dir { install }
 
-    assert_equal "Installed pre-commit hook.\n", @stdout.string
+    assert_equal "Installed post-commit hook.\n", @stdout.string
   end
 
   def test_should_refuse_a_hook_type_it_does_not_manage
@@ -76,7 +76,7 @@ class TestHookWriter < Minitest::Test
   def test_should_replace_a_hook_it_wrote_before
     in_git_dir { existing_hook("#!/bin/sh\n# fun-ci-managed-hook\nold-fun-ci-trigger\n") && install }
 
-    assert_equal FunCi::Setup::HookScript.for("pre-commit"), File.read(hook_path)
+    assert_equal FunCi::Setup::HookScript.for("post-commit"), File.read(hook_path)
   end
 
   private
@@ -86,14 +86,14 @@ class TestHookWriter < Minitest::Test
     yield
   end
 
-  def hook_path = File.join(@dir, ".git", "hooks", "pre-commit")
+  def hook_path = File.join(@dir, ".git", "hooks", "post-commit")
 
   def existing_hook(content)
     FileUtils.mkdir_p(File.dirname(hook_path))
     File.write(hook_path, content)
   end
 
-  def install(hook_type = "pre-commit")
+  def install(hook_type = "post-commit")
     FunCi::Setup::HookWriter.run(project_root: @dir, hook_type: hook_type, stdout: @stdout)
   end
 end